fix(color): Return `KeyColor` from `ColorPacket::add_key_color` if the packet is full
As `ColorPacket::add_key_color` takes ownership over the passed
`KeyColor`, it was consumed if the `ColorPacket` was already full.
In that case the caller could not add the `KeyColor` to a newly created
`ColorPacket`.

Now the error `KeyColorError::PacketFull` conatains the key_color and
therefore passes it back to the caller.