ref(utils): Add UsbWrapper containing all libusb references
As rust does not allow sibling structs, we had to get passed the Context
and DeviceHandle in order for the compiler to assure, that DeviceHandle
outlives Context and AsyncGroup outlives both.
Now we create Context, DeviceHandle und AsyncGroup internally and leak
them, to get &'static references to them. We group them in UsbWrapper,
which frees the memory when it is dropped.