fix(util): Don't unwrap during drop
Don't unwrap in the UsbWrappers handle method, not even if the thread is
not yet panicing. If the device is disconnected and a reconnect is
attempted, the old UsbWrapper is dropped and a new one will be created.
Dropping the old one with the device disconnected, this resulted in
unwrap() panicking.